powercap/intel_rapl: Use bitmap for Power Limits
the number of Power Limits per domain does not represent which Power
Limits are available.

Use bitmap to represent the availability of Power Limits.

Note that POWER_LIMIT1 is mandatory thus it does not need to be set
explicitly by RAPL Interface drivers.

+++ include/linux/intel_rapl.h
@@ -133,7 +133,7 @@ struct rapl_if_priv {
 	enum cpuhp_state pcap_rapl_online;
 	u64 reg_unit;
 	u64 regs[RAPL_DOMAIN_MAX][RAPL_DOMAIN_REG_MAX];
-	int limits[RAPL_DOMAIN_MAX];
+	unsigned long limits[RAPL_DOMAIN_MAX];
 	int (*read_raw)(int cpu, struct reg_action *ra);
 	int (*write_raw)(int cpu, struct reg_action *ra);
 	void *rpd;
